refactor: GradientBackground 파일명 카멜케이스로 변경
