+++ app/(beforeLogin)/_component/gradientBackground.tsx
