+++ app/(beforeLogin)/layout.tsx
@@ -1,5 +1,5 @@
 import {ReactNode} from "react";
-import GradientBackground from "@/app/(beforeLogin)/_component/GradientBackground";
+import GradientBackground from "@/app/(beforeLogin)/_component/gradientBackground";
 
 type Props = {children: ReactNode}
 export default function BeforeLoginLayout({children}: Props) {
